+++ test/src/codon_table.c
@@ -19,18 +19,18 @@ void test_codont_nonmarginal(void)
 
     struct nmm_codon* codon = nmm_codon_create(base);
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G')) == 0);
-    cass_cond(nmm_codon_lprob_set(codonp, codon, log(0.8)) == 0);
+    cass_cond(nmm_codon_lprob_set(codonp, codon, logf(0.8)) == 0);
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T')) == 0);
-    cass_cond(nmm_codon_lprob_set(codonp, codon, log(0.1)) == 0);
+    cass_cond(nmm_codon_lprob_set(codonp, codon, logf(0.1)) == 0);
 
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
     cass_cond(codont != NULL);
     nmm_codon_lprob_destroy(codonp);
 
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.8));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.8));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.1));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.1));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('T', 'T', 'T')) == 0);
     cass_cond(imm_lprob_is_zero(nmm_codon_table_lprob(codont, codon)));
     nmm_codon_destroy(codon);
@@ -46,30 +46,30 @@ void test_codont_marginal(void)
 
     struct nmm_codon* codon = nmm_codon_create(base);
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G'));
-    nmm_codon_lprob_set(codonp, codon, log(0.8));
+    nmm_codon_lprob_set(codonp, codon, logf(0.8));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T'));
-    nmm_codon_lprob_set(codonp, codon, log(0.1));
+    nmm_codon_lprob_set(codonp, codon, logf(0.1));
 
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
     cass_cond(codont != NULL);
     nmm_codon_lprob_destroy(codonp);
 
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.8));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.8));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.1));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.1));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'X')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.9));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.9));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'X', 'X')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.9));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.9));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('X', 'X', 'X')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.9));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.9));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('X', 'T', 'X')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.9));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.9));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('X', 'X', 'G')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.8));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.8));
     cass_cond(nmm_codon_set_triplet(codon, NMM_TRIPLET('X', 'X', 'T')) == 0);
-    cass_close(nmm_codon_table_lprob(codont, codon), log(0.1));
+    cass_close(nmm_codon_table_lprob(codont, codon), logf(0.1));
     nmm_codon_destroy(codon);
 
     nmm_codon_table_destroy(codont);
