+++ test/src/frame_state.c
@@ -25,14 +25,14 @@ void test_frame_state1(void)
     struct nmm_base_abc const*   base = nmm_base_abc_create("ACGT", 'X');
     struct imm_abc const*        abc = nmm_base_abc_super(base);
     struct nmm_base_table const* baset =
-        nmm_base_table_create(base, log(0.25), log(0.25), log(0.25), log(0.25));
+        nmm_base_table_create(base, logf(0.25), logf(0.25), logf(0.25), logf(0.25));
 
     struct nmm_codon_lprob* codonp = nmm_codon_lprob_create(base);
     struct nmm_codon*       codon = nmm_codon_create(base);
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G'));
-    nmm_codon_lprob_set(codonp, codon, log(0.8 / 0.9));
+    nmm_codon_lprob_set(codonp, codon, logf(0.8 / 0.9));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T'));
-    nmm_codon_lprob_set(codonp, codon, log(0.1 / 0.9));
+    nmm_codon_lprob_set(codonp, codon, logf(0.1 / 0.9));
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
     nmm_codon_lprob_destroy(codonp);
 
@@ -80,14 +80,14 @@ void test_frame_state2(void)
     struct nmm_base_abc const*   base = nmm_base_abc_create("ACGT", 'X');
     struct imm_abc const*        abc = nmm_base_abc_super(base);
     struct nmm_base_table const* baset =
-        nmm_base_table_create(base, log(0.1), log(0.2), log(0.3), log(0.4));
+        nmm_base_table_create(base, logf(0.1), logf(0.2), logf(0.3), logf(0.4));
 
     struct nmm_codon_lprob* codonp = nmm_codon_lprob_create(base);
     struct nmm_codon*       codon = nmm_codon_create(base);
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G'));
-    nmm_codon_lprob_set(codonp, codon, log(0.8 / 0.9));
+    nmm_codon_lprob_set(codonp, codon, logf(0.8 / 0.9));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T'));
-    nmm_codon_lprob_set(codonp, codon, log(0.1 / 0.9));
+    nmm_codon_lprob_set(codonp, codon, logf(0.1 / 0.9));
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
     nmm_codon_lprob_destroy(codonp);
 
@@ -150,16 +150,16 @@ void test_frame_state3(void)
     struct nmm_base_abc const*   base = nmm_base_abc_create("ACGT", 'X');
     struct imm_abc const*        abc = nmm_base_abc_super(base);
     struct nmm_base_table const* baset =
-        nmm_base_table_create(base, log(0.1), log(0.2), log(0.3), log(0.4));
+        nmm_base_table_create(base, logf(0.1), logf(0.2), logf(0.3), logf(0.4));
 
     struct nmm_codon_lprob* codonp = nmm_codon_lprob_create(base);
     struct nmm_codon*       codon = nmm_codon_create(base);
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G'));
-    nmm_codon_lprob_set(codonp, codon, log(0.8) - log(1.3));
+    nmm_codon_lprob_set(codonp, codon, logf(0.8) - logf(1.3));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T'));
-    nmm_codon_lprob_set(codonp, codon, log(0.1) - log(1.3));
+    nmm_codon_lprob_set(codonp, codon, logf(0.1) - logf(1.3));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('G', 'T', 'C'));
-    nmm_codon_lprob_set(codonp, codon, log(0.4) - log(1.3));
+    nmm_codon_lprob_set(codonp, codon, logf(0.4) - logf(1.3));
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
     nmm_codon_lprob_destroy(codonp);
 
@@ -227,7 +227,7 @@ void test_frame_state_lposterior(void)
     struct nmm_base_abc const*   base = nmm_base_abc_create("ACGT", 'X');
     struct imm_abc const*        abc = nmm_base_abc_super(base);
     struct nmm_base_table const* baset =
-        nmm_base_table_create(base, log(0.1), log(0.2), log(0.3), log(0.4));
+        nmm_base_table_create(base, logf(0.1), logf(0.2), logf(0.3), logf(0.4));
 
     char const* symbols = imm_abc_symbols(abc);
     unsigned    length = imm_abc_length(abc);
@@ -241,16 +241,16 @@ void test_frame_state_lposterior(void)
     while ((codon_item = cartes_next(codon_iter)) != NULL) {
         nmm_codon_set_triplet(codon,
                               (struct nmm_triplet){codon_item[0], codon_item[1], codon_item[2]});
-        nmm_codon_lprob_set(codonp, codon, log(0.001));
+        nmm_codon_lprob_set(codonp, codon, logf(0.001));
     }
     cartes_destroy(codon_iter);
 
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G'));
-    nmm_codon_lprob_set(codonp, codon, log(0.8));
+    nmm_codon_lprob_set(codonp, codon, logf(0.8));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T'));
-    nmm_codon_lprob_set(codonp, codon, log(0.1));
+    nmm_codon_lprob_set(codonp, codon, logf(0.1));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('G', 'T', 'C'));
-    nmm_codon_lprob_set(codonp, codon, log(0.4));
+    nmm_codon_lprob_set(codonp, codon, logf(0.4));
 
     cass_cond(nmm_codon_lprob_normalize(codonp) == 0);
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
@@ -264,7 +264,7 @@ void test_frame_state_lposterior(void)
         nmm_codon_set_triplet(codon,
                               (struct nmm_triplet){codon_item[0], codon_item[1], codon_item[2]});
 
-        double total = imm_lprob_zero();
+        float total = imm_lprob_zero();
         for (unsigned times = 1; times < 6; ++times) {
 
             struct cartes* seq_iter = cartes_create(symbols, length, times);
@@ -273,7 +273,7 @@ void test_frame_state_lposterior(void)
             while ((seq = cartes_next(seq_iter)) != NULL) {
                 struct imm_seq const* tmp = imm_seq_create(seq, abc);
                 cass_cond(tmp != NULL);
-                double lprob = nmm_frame_state_lposterior(state, codon, tmp);
+                float lprob = nmm_frame_state_lposterior(state, codon, tmp);
                 imm_seq_destroy(tmp);
                 lprob -= nmm_codon_table_lprob(codont, codon);
                 total = imm_lprob_add(total, lprob);
@@ -296,16 +296,16 @@ void test_frame_state_decode(void)
     struct nmm_base_abc const*   base = nmm_base_abc_create("ACGT", 'X');
     struct imm_abc const*        abc = nmm_base_abc_super(base);
     struct nmm_base_table const* baset =
-        nmm_base_table_create(base, log(0.1), log(0.2), log(0.3), log(0.4));
+        nmm_base_table_create(base, logf(0.1), logf(0.2), logf(0.3), logf(0.4));
 
     struct nmm_codon_lprob* codonp = nmm_codon_lprob_create(base);
     struct nmm_codon*       codon = nmm_codon_create(base);
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'G'));
-    nmm_codon_lprob_set(codonp, codon, log(0.8) - log(1.3));
+    nmm_codon_lprob_set(codonp, codon, logf(0.8) - logf(1.3));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('A', 'T', 'T'));
-    nmm_codon_lprob_set(codonp, codon, log(0.1) - log(1.3));
+    nmm_codon_lprob_set(codonp, codon, logf(0.1) - logf(1.3));
     nmm_codon_set_triplet(codon, NMM_TRIPLET('G', 'T', 'C'));
-    nmm_codon_lprob_set(codonp, codon, log(0.4) - log(1.3));
+    nmm_codon_lprob_set(codonp, codon, logf(0.4) - logf(1.3));
     struct nmm_codon_table const* codont = nmm_codon_table_create(codonp);
     nmm_codon_lprob_destroy(codonp);
